test(interceptor): split server cli tests by feature
